+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to
 
 ### Added
 
+- Added discount column to csv order export
 - Added actions in admin api endpoints `BatchOrder` to validate payment,
   generate orders and send voucher codes, and send invitation signature link
 - Add admin api endpoints for `BatchOrder` resources to create, read, list